% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.R
\name{BasicHeatmap}
\alias{BasicHeatmap}
\title{Draw Basic heatmaps with not annotation}
\usage{
BasicHeatmap(
  p.data,
  color_min = "cyan",
  color_mid = "black",
  color_max = "gold",
  break_min = -4,
  break_mid = 0,
  break_max = 4,
  cluster_type = "hclust",
  cluster_row = FALSE,
  cluster_col = FALSE,
  km_col = 3,
  km_row = 3,
  km_repeat_time = 100,
  cluster_row_method = "complete",
  cluster_row_distance = "euclidean",
  cluster_col_method = "complete",
  cluster_col_distance = "euclidean",
  show_rowname = T,
  show_colname = T
)
}
\arguments{
\item{p.data}{normlized matrix}

\item{color_min}{Color for minmal values.}

\item{color_mid}{Color for middle values.}

\item{color_max}{Color for max values.}

\item{break_min}{min values.}

\item{break_mid}{mid values.}

\item{break_max}{max values.}

\item{cluster_type}{Hierarchical clustering or k-means.}

\item{cluster_row}{For hclust, cluster row or not, default = FALSE.}

\item{cluster_col}{For hclust, cluster col or not, default = FALSE.}

\item{km_col}{For k-means, number of expected group of column.}

\item{km_row}{For k-means, number of expected group of column.}

\item{km_repeat_time}{Avoid different results of k-means each time, Default = 100.}

\item{cluster_row_method}{see hclust.}

\item{cluster_row_distance}{see hclust.}

\item{cluster_col_method}{see hclust.}

\item{cluster_col_distance}{see hclust.}

\item{show_rowname}{Show rownames or not.}

\item{show_colname}{Show colnames or not.}
}
\value{
a CompexHeatmap result
}
\description{
Draw Basic heatmaps with not annotation
}
\references{
https://jokergoo.github.io/ComplexHeatmap-reference/book/
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>
}
