% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.R
\name{Datacleaning}
\alias{Datacleaning}
\title{Functions for data cleaning, aiming at remove na values or replace it.}
\usage{
Datacleaning(
  x,
  na_value_treat = "fill",
  na_tag = "NA",
  na_replace_value,
  remove_noise = FALSE,
  sample_percetage,
  noise_cutoff
)
}
\arguments{
\item{x}{Data for heatmap.}

\item{na_value_treat}{How to deal with NA values, default is "fill" which means you have to provide a value to replace NA value eg:0.}

\item{na_tag}{If NA in the original data is not in the form of blank, but in the form of other characters such as "N/F", it needs to be marked.}

\item{na_replace_value}{The value will replace the NA value.}

\item{remove_noise}{For example, there may be noise in transcriptome data and metabolome data, and the noise needs to be removed before visualization}

\item{sample_percetage}{How to remove noise: where the value is greater than the threshold "noise_cutoff" in "sample_percetage" percent of the samples.}

\item{noise_cutoff}{How to remove noise: where the value is greater than the threshold "noise_cutoff" in "sample_percetage" percent of the samples.}
}
\value{
dx a matrix like datExpr
}
\description{
Functions for data cleaning, aiming at remove na values or replace it.
}
\references{
https://jokergoo.github.io/ComplexHeatmap-reference/book/
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>
}
