% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.R
\name{getFinalHeatmap}
\alias{getFinalHeatmap}
\title{Add annotations}
\usage{
getFinalHeatmap(
  basicPlot,
  row_anno_side = "left",
  col_anno_side = "top",
  anno_row,
  anno_col
)
}
\arguments{
\item{basicPlot}{Heatmap generate by 1st step.}

\item{row_anno_side}{Location of row annotation}

\item{col_anno_side}{Location of column annotation}

\item{anno_row}{Need row annotation or not?}

\item{anno_col}{Need col annotation or not?}
}
\value{
will generate final heatmap plot
}
\description{
Add annotations
}
\references{
https://jokergoo.github.io/ComplexHeatmap-reference/book/
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>
}
