% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.R
\name{getNormTable}
\alias{getNormTable}
\title{Draw Basic heatmaps with not annotation}
\usage{
getNormTable(x, normlize, scale)
}
\arguments{
\item{x}{Data for heatmap after cleaning.}

\item{normlize}{Normlization method for log2,log10,log2 x+1 ,log10 x+1}

\item{scale}{Z-score scaled data input, rowscale, colscale and none, Default = "row".}
}
\value{
a Normlized heatmap data
}
\description{
Draw Basic heatmaps with not annotation
}
\references{
https://jokergoo.github.io/ComplexHeatmap-reference/book/
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>
}
