% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFun.R
\name{HeatAnno}
\alias{HeatAnno}
\title{generate annotations for row and column}
\usage{
HeatAnno(anno_row = NULL, anno_col = NULL)
}
\arguments{
\item{anno_row}{annotations for row}

\item{anno_col}{annotations for column.}
}
\value{
A list contains row_annotation and col_annotatin df will generated.
}
\description{
generate annotations for row and column
}
\references{
https://jokergoo.github.io/ComplexHeatmap-reference/book/
}
\author{
Shawn Wang <http://www.shawnlearnbioinfo.top>
}
